added basic styling
