+++ ui/.env
@@ -1,0 +1,1 @@
+VITE_GEONAMES_USERNAMES=whalefan
