+++ ui/src/api/ApiClient.ts
@@ -95,3 +95,4 @@ export async function login(email: string, password: string) {
     
     return response.json();
 }
+
