+++ ui/src/pages/Home/HomePage.tsx
@@ -1,8 +1,11 @@
+import { Statistics } from "../../components/Statistics/Statistics";
+
 const HomePage = () => {
     
     return (
         <div>
             <h1 className="title">Test Home Page</h1>
+              <Statistics />
         </div>
         );
 }
